DRAFT: Feature: Non-privileged hacluster
It enables creating the cluster under the hacluster
It requires some preparaions. To use it, do:

Bring the user to the haclient group. Assume the user is 'hackerman'

$ user=hackerman
$ sudo sed -i "s|$user:x:1000:1000|$user:x:1000:90|g" /etc/passwd
$ reboot -h now

Then do

$ user=hackerman
$ sudo ln -s /sbin/crm* /sbin/cibadmin /sbin/stonith /sbin/corosync-keygen \
        /sbin/corosync-cfgtool /sbin/csync2 /sbin/sbd /usr/bin/ \
        /usr/bin/

$ sudo chmod 770 /etc/pacemaker/
$ sudo chown $user:haclient -R /etc/csync2 /etc/corosync \
        /usr/share/doc/packages/corosync \
        /etc/sysconfig /var/lib/csync2 /etc/crm /etc/lvm /etc/samba
$ sudo chmod 777 -R /etc/csync2 /etc/corosync \
        /usr/share/doc/packages/corosync \
        /etc/sysconfig /var/lib/csync2 /etc/crm /etc/lvm /etc/samba
ATTENTION: /run resets its permissions on the node restart
$ sudo chown $user:haclient /run
$ sudo chmod 777 /run
$ sudo chmod 777 /var/log

Besides set up the csync2 running as hacluster
Add 'User=hacluster' into the '[Service]' section
$ sudo echo "User=$user" >> /usr/lib/systemd/system/csync2@.service
$ sudo systemctl daemon-reload

If you also want it working with -N key, the hacluster should be in
/etc/sudoers
$ sudo echo "$user ALL=(ALL:ALL) NOPASSWD: ALL" >> /etc/sudoers

And now you can do the usual routines under hacluster

$ crm cluster init ...
$ crm cluster join ...
  or
$ crm cluster init -N node1 -N node2 ..