Dev: crmcheck: (new feature) call trento to do checks
This commit adds two new commands

$ crm check list
$ crm check execute <check-id1> [check-id2|...]

How it works?
It gets the trento-wand key from the crm.conf and sends the rest-api request
together with the key to the running wanda (wanda address is in crm.conf).
Then it receives the response and shows the result.

If there is no wanda key, it gets the credentials from the crm.conf,
send them to the trento-web (web address is in crm.conf), and
it gets the wanda key in return. The rest is the same.